[KYUUBI #1565] Move time functions to RowSetUtils and move schema to spark package
